refactor: Ajuste na orgização de arquivos e na estrutura do projeto para melhor escalabilidade
