+++ projeto-map/myapp/api/v1/viewsets/authViewSet.py
@@ -1,16 +1,15 @@
 # myapp/api/v1/viewsets.py
 
 from django.contrib.auth import authenticate
-
 from rest_framework import status, viewsets
 from rest_framework.decorators import action
 from rest_framework.permissions import AllowAny, IsAuthenticated
 from rest_framework.response import Response
 from rest_framework_simplejwt.tokens import RefreshToken
-from .serializers.RegistrationSerializer import (UserRegistrationSerializer)
-from .serializers.LoginSerializer import(UserLoginSerializer)
-from .serializers.ProfileSerializer import(UserProfileSerializer)
-from .serializers.ChangePasswordSerializer import(ChangePasswordSerializer)
+from ..serializers.RegistrationSerializer import (UserRegistrationSerializer)
+from ..serializers.LoginSerializer import(UserLoginSerializer)
+from ..serializers.ProfileSerializer import(UserProfileSerializer)
+from ..serializers.ChangePasswordSerializer import(ChangePasswordSerializer)
