+++ projeto-map/myproject/router/api_v1.py
@@ -1,7 +1,7 @@
 from django.conf.urls import include
 from django.urls import path
 
-from myapp.api.v1.router import router
+from myapp.api.v1.routes.authRouter import router
 
 api_v1_urls = [
     path("", include((router.urls, "myapp"), namespace="myapp")),
