+++ projeto-map/requirements.txt
@@ -15,3 +15,4 @@ rpds-py==0.23.1
 sqlparse==0.5.3
 typing_extensions==4.12.2
 uritemplate==4.1.1
+psycopg2-binary==2.9.9
